energy/requirements.txt
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.10,<0.1.0
langchain-core>=0.1.10,<0.2.0
openai>=1.3.7
pandas>=2.1.4
numpy>=1.26.2
chromadb>=0.4.22
python-dotenv>=1.0.0
tiktoken>=0.5.2
streamlit>=1.29.0
plotly>=5.18.0
